+++ marshmallow/fields.py
@@ -719,7 +719,7 @@ class Number(Field):
             return self.default
         try:
             return self._format_num(value)
-        except (TypeError, ValueError, decimal.InvalidOperation) as err:
+        except (TypeError, ValueError) as err:
             raise exception_class(getattr(self, 'error', None) or err)
 
     def serialize(self, attr, obj, accessor=None):
@@ -785,11 +785,21 @@ class Decimal(Number):
         self.rounding = rounding
         super(Decimal, self).__init__(default=default, as_string=as_string, **kwargs)
 
+    # override Number
     def _format_num(self, value):
         num = decimal.Decimal(value)
         if self.places is not None:
             num = num.quantize(self.places, rounding=self.rounding)
         return num
+
+    # override Number
+    def _validated(self, value, exception_class):
+        try:
+            return super(Decimal, self)._validated(value, exception_class)
+        except decimal.InvalidOperation:
+            raise exception_class(
+                getattr(self, 'error', None) or 'Invalid decimal value.'
+            )
 
 
 class Boolean(Field):
@@ -813,6 +823,7 @@ class Boolean(Field):
         try:
             value_str = text_type(value)
         except TypeError as error:
+            msg = getattr(self, 'error', None) or text_type(error)
             raise UnmarshallingError(error)
         if value_str in self.falsy:
             return False
@@ -820,10 +831,11 @@ class Boolean(Field):
             if value_str in self.truthy:
                 return True
             else:
-                raise UnmarshallingError(
-                    '{0!r} is not in {1} nor {2}'.format(
-                        value_str, self.truthy, self.falsy
-                    ))
+                default_message = '{0!r} is not in {1} nor {2}'.format(
+                    value_str, self.truthy, self.falsy
+                )
+                msg = getattr(self, 'error', None) or default_message
+                raise UnmarshallingError(msg)
         return True
 
 class FormattedString(Field):